Add role selection API callbacks and middleware
Added role selection callbacks in order to set the role inside the table row of the user which is necessary for web app functionality.
Did that by creating a new view that handles GET and POST requests regarding the user role and also by creating a middleware that ensures the correct page redirection if the user doesn't have a role.